src/ingest_json/requirements.txt
google-cloud-bigquery~=3.27
google-cloud-storage~=2.19